fix setup script with --allow-new
Until now, I had tested the setup script with my own config which
includes `git.push-new-bookmarks = true`. Because of that, I didn't
notice the bug caused by the missing --allow-new flag.

So I'm adding it to the script just so I can keep working, but I can't
publish the tutorial before a better UI is added to Jujutsu that I can
actually explain to readers.