replace `describe --reset-author` with `metaedit --update-author`
This is a little early, as it's a breaking change for anyone that
doesn't have the latest version of Jujutsu, which was released just a
couple hours ago. But there is a version check in the reset script now,
so people should quickly notice the root of the problem if some commands
don't work right. It's much easier for me to be able to rely on the
latest features.