use normal sections for faq
It's not immediately clear that the text boxes are expandable. The next
time a text box appears, it is explained that it's expandable.

I used the collapsed boxes because I thought it would communicate that
these sections are optional. But I guess it should be clear from the
section titles, too. If the reader isn't interested in the question,
they'll probably skip it on their own.