clarify installation of text editor
Previously, it wasn't clear that there is a default editor that works
fine and the whole section is optional. There also was no like to the
repo of `edit`, so it was difficult for people to figure out how to
install it if they didn't use `mise` to install Jujutsu.